Added option to not delete data on uninstall
git-svn-id: https://plugins.svn.wordpress.org/contact-form-7-to-database-extension/trunk@365398 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ CF7DBPlugin.php
@@ -63,6 +63,7 @@ class CF7DBPlugin extends CF7DBPluginLifeCycle {
             'SaveCookieData' => array(__('Save Cookie Data with Form Submissions', 'contact-form-7-to-database-extension'), 'false', 'true'),
             'SaveCookieNames' => array(__('Save only cookies in DB named (comma-separated list, no spaces, and above option must be set to true)', 'contact-form-7-to-database-extension')),
             'ShowQuery' => array(__('Show the query used to display results', 'contact-form-7-to-database-extension'), 'false', 'true'),
+            'DropOnUninstall' => array(__('Drop this plugin\'s Database table on uninstall', 'contact-form-7-to-database-extension'), 'true', 'false'),
             //'SubmitTableNameOverride' => array(__('Use this table to store submission data rather than the default (leave blank for default)', 'contact-form-7-to-database-extension'))
         );
     }
@@ -185,14 +186,16 @@ class CF7DBPlugin extends CF7DBPluginLifeCycle {
      * @return void
      */
     protected function unInstallDatabaseTables() {
-        global $wpdb;
-        $tableName = $this->getSubmitsTableName();
-        $wpdb->query("DROP TABLE IF EXISTS $tableName");
-        //        $tables = array('SUBMITS');
-        //        foreach ($tables as $aTable) {
-        //            $tableName = $this->prefixTableName($aTable);
-        //            $wpdb->query("DROP TABLE IF EXISTS $tableName");
-        //        }
+        if ('true' == $this->getOption('DropOnUninstall', 'true')) {
+            global $wpdb;
+            $tableName = $this->getSubmitsTableName();
+            $wpdb->query("DROP TABLE IF EXISTS $tableName");
+            //        $tables = array('SUBMITS');
+            //        foreach ($tables as $aTable) {
+            //            $tableName = $this->prefixTableName($aTable);
+            //            $wpdb->query("DROP TABLE IF EXISTS $tableName");
+            //        }
+        }
     }
 
     public function addActionsAndFilters() {
